mizerSeasonal: Extends the mizer package to allow for seasonal reproduction

This is an extension package for the mizer package

See also

Author

Maintainer: Gustav Delius gustav.delius@york.ac.uk (ORCID) [copyright holder]

Authors:

Michael A. Spence michael.spence@cefas.gov.uk (ORCID) [copyright holder]