Beverton Holt function to calculate density-dependent reproduction rate

Takes the density-independent rates \(R_{di}(t)\) of egg production and returns reduced, density-dependent reproduction rates \(R_{dd}(t)\) given as $$R_{dd}(t) = R_{di}(t) \frac{R_{max}(t)}{R_{di}(t) + R_{max}(t)}$$ where \(R_{max}(t)\) are the maximum possible reproduction rates that must be specified in an array (time x species) saved in the other_params$R_max slot of the MizerParams object. This is simply a time-dependent version of mizer::BevertonHoltRDD(). (All quantities in the above equation are species-specific but we dropped the species index for simplicity.)

Usage

seasonalBevertonHoltRDD(rdi, params, t, ...)

Arguments

rdi

Vector of density-independent reproduction rates \(R_{di}\) for all species.

params

A params object that must contain other_params$R_max

t

The time at which to calculate RDD

...

Unused

Value

Vector of density-dependent reproduction rates.

See also