Set pulsed reproduction

This returns a new model in which the reproduction is set to take place only within a single time step each year. During the rest of the year the energy available for reproduction is accumulated as gonadic mass, all of which is realeased during the time step into which the pulse_time falls.

setPulsedReproduction(params, pulse_time)

Arguments

params

A MizerParams object

pulse_time

A number between 0 and 1 that specifies the time within each year at which reproduction should take place.

Value

A MizerParams object with the updated model